Move 'get-parent-class-name' out of library
