+++ src/cem-analyzer-plugins/add-events.ts
@@ -10,7 +10,7 @@ import {
   SyntaxKind,
   type Node,
 } from 'typescript';
-import getParentClassName from '../library/get-parent-class-name.js';
+import getParentClassName from './get-parent-class-name.js';
 
 /**
  * Adds dispatched events to the manifest.
